mudando para int, modificando nomes variaveis

+++ src/Escalonador.java
@@ -4,17 +4,17 @@ import java.util.Queue;
 public class Escalonador {
 
 	private Queue<Fregues> fila;
-	private double tempoMedioAtendendo;
+	private int tempoMedioAtendendo;
 	private int reqAtendidas;
 	private Servidor servidor;
 	private Distribuicao dist;
 
-	public Escalonador(double tempoMedioServico, Servidor servidor) {
+	public Escalonador(int tempoMedioServico, Servidor servidor) {
 		this.fila = new PriorityQueue<Fregues>();
 		this.tempoMedioAtendendo = 0;
 		this.reqAtendidas = 0;
 		this.servidor = servidor;
-		double[] params = new double[1];
+		int[] params = new int[1];
 		params[0] = tempoMedioServico;
 		this.dist = new Distribuicao(TipoDistribuicao.EXPONENCIAL, params);
 	}
@@ -34,10 +34,10 @@ public class Escalonador {
 	
 	public double escalonar(Fregues fregues){
 		this.servidor.ocupar();
-		double tempoTermino = this.dist.gerar();
-		fregues.finalizar(tempoTermino + System.currentTimeMillis());
-		double tempoAtendido = fregues.getTempoAtendido();
-		this.tempoMedioAtendendo = (this.tempoMedioAtendendo * (this.reqAtendidas - 1) + tempoAtendido) / (this.reqAtendidas + 1);
+		int tempoTermino = this.dist.gerar();
+		fregues.finalizar((int) (tempoTermino + System.currentTimeMillis()));
+		int tempoAtendido = fregues.getTempoAtendido();
+		this.tempoMedioAtendendo = (int) ((this.tempoMedioAtendendo * (this.reqAtendidas - 1) + tempoAtendido) / (this.reqAtendidas + 1));
 		return tempoTermino + System.currentTimeMillis();
 	}
 }
